Switch default RNG implementation from UniversalGenerator to MarsenneTwister
CAREFUL - Universal Generator is bugged and returned ints via nextInt()
will have least significant bits either all zeros or ones.

+++ math/src/main/java/smile/math/random/MersenneTwister.java
@@ -69,7 +69,7 @@ public class MersenneTwister implements RandomNumberGenerator {
     /**
      * Constructor.
      */
-    public MersenneTwister(int seed) {
+    public MersenneTwister(long seed) {
         setSeed(seed);
     }
